Break circular require()'s
