+++ Extension/src/main.ts
@@ -67,7 +67,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<CppToo
     // Notify users if debugging may not be supported on their OS.
     util.checkDistro(info);
     await checkVsixCompatibility();
-    UpdateInsidersAccess();
+    LanguageServer.UpdateInsidersAccess();
 
     const settings: CppSettings = new CppSettings((vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length > 0) ? vscode.workspace.workspaceFolders[0]?.uri : undefined);
     let isOldMacOs: boolean = false;
@@ -212,45 +212,6 @@ function sendTelemetry(info: PlatformInformation): void {
             break;
     }
     Telemetry.logDebuggerEvent("acquisition", telemetryProperties);
-}
-
-export function UpdateInsidersAccess(): void {
-    let installPrerelease: boolean = false;
-
-    // Only move them to the new prerelease mechanism if using updateChannel of Insiders.
-    const settings: CppSettings = new CppSettings();
-    const migratedInsiders: PersistentState<boolean> = new PersistentState<boolean>("CPP.migratedInsiders", false);
-    if (settings.updateChannel === "Insiders") {
-        // Don't do anything while the user has autoUpdate disabled, so we do not cause the extension to be updated.
-        if (!migratedInsiders.Value && vscode.workspace.getConfiguration("extensions", null).get<boolean>("autoUpdate")) {
-            installPrerelease = true;
-            migratedInsiders.Value = true;
-        }
-    } else {
-        // Reset persistent value, so we register again if they switch to "Insiders" again.
-        if (migratedInsiders.Value) {
-            migratedInsiders.Value = false;
-        }
-    }
-
-    // Mitigate an issue with VS Code not recognizing a programmatically installed VSIX as Prerelease.
-    // If using VS Code Insiders, and updateChannel is not explicitly set, default to Prerelease.
-    // Only do this once. If the user manually switches to Release, we don't want to switch them back to Prerelease again.
-    if (util.isVsCodeInsiders()) {
-        const insidersMitigationDone: PersistentState<boolean> = new PersistentState<boolean>("CPP.insidersMitigationDone", false);
-        if (!insidersMitigationDone.Value) {
-            if (vscode.workspace.getConfiguration("extensions", null).get<boolean>("autoUpdate")) {
-                if (settings.getWithUndefinedDefault<string>("updateChannel") === undefined) {
-                    installPrerelease = true;
-                }
-            }
-            insidersMitigationDone.Value = true;
-        }
-    }
-
-    if (installPrerelease) {
-        vscode.commands.executeCommand("workbench.extensions.installExtension", "ms-vscode.cpptools", { installPreReleaseVersion: true });
-    }
 }
 
 async function checkVsixCompatibility(): Promise<void> {
